% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/wrappers.R
\name{CalcSWTPower}
\alias{CalcSWTPower}
\title{GEE-based power calculation for SWTs and MLIs}
\usage{
CalcSWTPower(
  design_mat_list = NULL,
  dgm_param_vec = NULL,
  n_clust_per_seq = NULL,
  n_ind_per_clust = NULL,
  crossover_time_period_vec = NULL,
  clust_trt_effect = NULL,
  ind_trt_effect = NULL,
  interaction_effect = NULL,
  period_effect_vec = NULL,
  incr_effect_scalar = NULL,
  var_fun,
  cor_mat_list = NULL,
  incidence_mat_list = NULL,
  dist = "normal",
  link = "identity",
  dispersion_scalar_scalar = NULL,
  alpha = 0.05,
  test_df,
  contrast_mat = NULL
)
}
\description{
Perform the GEE-based power calculation for stepped wedge trials
and multilevel cluster randomized trials.
}
