% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_design_matrices.R
\name{CreateClusterCompleteDesignMatrixList}
\alias{CreateClusterCompleteDesignMatrixList}
\title{Create Cluster Complete Design Matrix List}
\usage{
CreateClusterCompleteDesignMatrixList(
  n_study_periods,
  n_clust_trt_seqs,
  n_clust_per_seq,
  crossover_time_period_vec,
  n_ind_per_clust,
  time_model_type,
  trt_model_type,
  linear_trt_scale_factor = 1,
  mli_study_flag = FALSE
)
}
\arguments{
\item{n_study_periods}{Integer. The number of study periods.}

\item{n_clust_trt_seqs}{Integer. The number of cluster treatment sequences.}

\item{n_clust_per_seq}{Integer. The number of clusters per sequence.}

\item{crossover_time_period_vec}{Numeric vector. The time periods for crossover.}

\item{n_ind_per_clust}{Integer. The number of individuals per cluster.}

\item{time_model_type}{Character. The type of time model ("linear" or "categorical").}

\item{trt_model_type}{Character. The type of treatment model ("average" or "linear").}

\item{linear_trt_scale_factor}{Numeric. The scaling factor for linear treatment effects. Default is 1.}

\item{mli_study_flag}{Logical. Flag indicating if the study is a multi-level intervention study. Default is FALSE.}
}
\value{
A list of design matrices.
}
\description{
Creates a list of design matrices for multiple clusters over a given number of study periods.
These matrices account for the study design, treatment sequences, time effects, and other relevant factors.
}
