% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_incidence_matrices.R
\name{CreateSequenceIncidenceMat}
\alias{CreateSequenceIncidenceMat}
\title{Generate the incidence matrix for a specific sequence (applies to all clusters in that sequence)}
\usage{
CreateSequenceIncidenceMat(
  start_period,
  end_period,
  n_study_periods,
  unobserved_periods = NULL
)
}
\arguments{
\item{start_period}{The first time period in which the cluster is observed}

\item{end_period}{The last time period in which the cluster is observed}

\item{n_study_periods}{The total number of time periods in the study.}

\item{unobserved_periods}{An optional vector of time periods in which the cluster is not observed. Only necessary if the cluster is not observed in time periods between the start and end period.
Not required if unobserved periods are at the beginning or end of the study.}
}
\value{
A matrix with one row for each time period the cluster is observed and .
}
\description{
r lifecycle::badge("experimental")`
The incidence matrix defines the time periods in which each cluster is observed.
Each row corresponds to a time period the cluster is observed, and each column corresponds to a time period in the study.
}
\examples{
CreateIncidenceMat(start_period = 1,
                  end_period = 3,
                 n_study_periods = 5)

CreateIncidenceMat(start_period = 3,
                 end_period = 7,
                n_study_periods = 10,
               unobserved_periods = c(4, 5))

}
