% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_power.R
\name{.CalcEstimatedParamVec}
\alias{.CalcEstimatedParamVec}
\title{Calculate the Estimated Parameter Vector (\( \hat{\theta} \))}
\usage{
.CalcEstimatedParamVec(
  design_mat_list,
  mean_jacobian_mat_list,
  var_mat_vi_list,
  mui_vec_list
)
}
\arguments{
\item{design_mat_list}{A list of numeric matrices representing the design matrices \( X_i \) for each cluster.}

\item{mean_jacobian_mat_list}{A list of numeric matrices representing the mean Jacobian matrices \( D_i \) for each cluster.}

\item{var_mat_vi_list}{A list of numeric matrices representing the covariance matrices \( V_i \) for each cluster.}

\item{mui_vec_list}{A list of numeric vectors representing the expected responses \( \mu_i \) for each cluster.}
}
\value{
Returns a numeric vector \( \hat{\theta} \), the estimated parameter vector.
}
\description{
This internal function calculates the estimated parameter vector \( \hat{\theta} \)
based on the design matrices, mean Jacobian matrices, covariance matrices, expected responses,
and the inverse link function.
}
\details{
The estimated parameter vector \( \hat{\theta} \) is calculated as:
[
\hat{\theta} = \left( \sum_{i = 1}^I X_i^T D_i V_i D_i^T X_i \right)^{-1} \sum_{i = 1}^I X_i^T D_i V_i D_i^T g(\mu_i)
]
where \( X_i \) are the design matrices, \( D_i \) are the mean Jacobian matrices, \( V_i \) are the covariance matrices,
and \( g(\mu_i) \) are the expected responses transformed by the inverse link function.
}
\keyword{internal}
