% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_power.R
\name{.CalcPower}
\alias{.CalcPower}
\title{Internal Power Calculation Function}
\usage{
.CalcPower(
  design_mat_list,
  working_cor_mat_list,
  incidence_mat_list = NULL,
  dgm_param_col_vec,
  dispersion_scalar,
  family,
  alpha,
  test_df,
  contrast_mat,
  null_val_vec,
  power_only_flag = FALSE
)
}
\arguments{
\item{design_mat_list}{A list of design matrices, one for each cluster.}

\item{working_cor_mat_list}{A list of working correlation matrices, one for each cluster.}

\item{incidence_mat_list}{A list of incidence matrices, one for each cluster.}

\item{dgm_param_col_vec}{A column vector of all data generating model parameters.}

\item{dispersion_scalar}{The dispersion parameter for the GEE model.}

\item{family}{A family object (from base R \code{stats} class)}

\item{alpha}{The desired Type I error rate.}

\item{test_df}{The degrees of freedom for the test statistic.}

\item{contrast_mat}{A matrix specifying the contrasts for hypothesis testing.}

\item{null_val_vec}{A vector specifying the null hypothesis values for the contrasts.}

\item{power_only_flag}{A logical flag indicating whether to return only the calculated power.
Used internally for sample size calculations.}
}
\value{
The function returns a list with the following components:
\itemize{
\item \code{power}: The power of the test.
\item \code{cluster_sample_size}: The sample size for each cluster.
\item \code{test_stat}: The test statistic.
\item \code{alpha}: The Type I error rate.
\item \code{test_df}: The degrees of freedom for the test statistic.
}
}
\description{
The function \code{.CalcPower} performs the core power calculation for stepped wedge trials and
multilevel cluster randomized trials using Generalized Estimating Equations (GEE).
This function is not intended to be directly called by the user, but rather serves as
an internal function for the main power calculation function.
}
\details{
The function performs the following calculations:
\enumerate{
\item Calculates the components for the estimating equation for each cluster: \( \boldsymbol{\mu}\emph{i \),
\( \textbf{D}}{i} \), \( \textbf{B}_{i} \), \( \textbf{R}\emph{i(\boldsymbol{\alpha}) \), and \( \textbf{V}}{i} \).
\item Solves for \( \boldsymbol{\Sigma}_1^{-1} \) and \( \widehat{\boldsymbol{\theta}} \).
\item Uses the Wald test to calculate the power based on the noncentrality parameter \( \lambda \),
the critical value, and the degrees of freedom \( q \).
}
}
\keyword{internal}
