% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_power.R
\name{.CalcVarMatrixVi}
\alias{.CalcVarMatrixVi}
\title{Calculate the Covariance Matrix (\( V_i \))}
\usage{
.CalcVarMatrixVi(var_mat_bi, working_cor_mat_ri, dispersion_scalar)
}
\arguments{
\item{var_mat_bi}{A numeric matrix representing the variance parameter matrix \( B_i \).}

\item{working_cor_mat_ri}{A numeric matrix representing the working correlation matrix \( R_i \).}
}
\value{
Returns a numeric matrix \( V_i \), the covariance matrix.
}
\description{
This internal function calculates the covariance matrix \( V_i \) based on the variance parameter matrix \( B_i \)
and the working correlation matrix \( R_i \).
}
\details{
The covariance matrix \( V_i \) is calculated as:
[
V_i = $\phi$ B_i^{1/2} R_i B_i^{1/2}
]
}
\keyword{internal}
