% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_power.R
\name{.CalcWaldNcp}
\alias{.CalcWaldNcp}
\title{Calculate the noncentrality parameter for the Wald test}
\usage{
.CalcWaldNcp(
  n_clust,
  contrast_mat,
  null_val_vec,
  estimated_param_vec,
  model_based_var_mat
)
}
\arguments{
\item{n_clust}{An integer representing the number of clusters.}

\item{contrast_mat}{A numeric matrix representing the contrast matrix \( L \).}

\item{null_val_vec}{A numeric vector representing the null value vector \( \ell \).}

\item{estimated_param_vec}{A numeric vector representing the estimated parameter vector \( \hat{\theta} \).}

\item{model_based_var_mat}{A numeric matrix representing the model-based variance matrix \( \Sigma_1 \).}
}
\value{
Returns a numeric value representing the noncentrality parameter \( \lambda \).
}
\description{
This internal function calculates the noncentrality parameter \( \lambda \) for the Wald test.
}
\details{
The noncentrality parameter \( \lambda \) is calculated as:
[
\lambda = I \left( L \hat{\theta} - \ell \right)^T \Sigma_1^{-1} \left( L \hat{\theta} - \ell \right)
]
}
\keyword{internal}
