% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_power.R
\name{.CalcWaldNullCritVal}
\alias{.CalcWaldNullCritVal}
\title{Calculate the Critical Value under the Null Hypothesis for the Wald Test}
\usage{
.CalcWaldNullCritVal(alpha, test_df)
}
\arguments{
\item{alpha}{A numeric value representing the type I error rate \( \alpha \).}

\item{test_df}{An integer representing the degrees of freedom.}
}
\value{
Returns a numeric value representing the critical value under the null hypothesis.
}
\description{
This internal function calculates the critical value under the null hypothesis for the Wald test.
}
\details{
The critical value under the null hypothesis is calculated using the chi-squared distribution quantile function:
[
\chi^2_{1-\alpha, \text{df}}
]
}
\keyword{internal}
