% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_design_matrices.R
\name{.CheckClusterCompleteDesignMatrixInputs}
\alias{.CheckClusterCompleteDesignMatrixInputs}
\title{Check Inputs for CreateClusterCompleteDesignMatrixList}
\usage{
.CheckClusterCompleteDesignMatrixInputs(
  n_study_periods,
  n_clust_trt_seqs,
  n_clust_per_seq,
  crossover_time_period_vec,
  n_ind_per_clust,
  time_model_type,
  trt_model_type,
  linear_trt_scale_factor = 1,
  mli_study_flag = FALSE
)
}
\value{
NULL. The function performs checks and will return an error if any check fails.
}
\description{
This function checks the validity of the inputs for the \code{CreateClusterCompleteDesignMatrixList} function.
It uses the \code{checkmate} package to perform these checks.
}
\details{
For more information on the parameters, see \code{\link{CreateClusterCompleteDesignMatrixList}}.
}
\keyword{internal,check}
