% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_design_matrices.R
\name{.CheckClusterCompleteDesignMatrixOutput}
\alias{.CheckClusterCompleteDesignMatrixOutput}
\title{Check Output for CreateClusterCompleteDesignMatrixList}
\usage{
.CheckClusterCompleteDesignMatrixOutput(design_mat_list)
}
\value{
NULL. The function performs checks and will return an error if any check fails.
}
\description{
This function checks the validity of the output for the \code{CreateClusterCompleteDesignMatrixList} function.
Specifically, it checks the number of columns in each matrix in the list to ensure they are consistent.
}
\details{
For more information on the expected output structure, see \code{\link{CreateClusterCompleteDesignMatrixList}}.
}
\keyword{internal,check}
