% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mean_model_param_vector.R
\name{.CheckDGMInputs}
\alias{.CheckDGMInputs}
\title{Check Inputs for Data Generating Model Specification}
\usage{
.CheckDGMInputs(
  n_study_periods,
  mli_study_flag,
  time_model_type,
  time_intercept_param,
  time_trend_param,
  period_effect_param_vec,
  trt_model_type,
  clust_trt_param,
  ind_trt_param,
  interaction_param
)
}
\value{
NULL. Function returns NULL if all checks pass.
}
\description{
Internal function to validate the inputs for the \code{SpecifyDataGeneratingModel} function.
For details on the parameters, see the documentation for \code{SpecifyDataGeneratingModel}.
}
\seealso{
\code{\link{SpecifyDataGeneratingModel}}
}
\keyword{internal,check}
