% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mean_model_param_vector.R
\name{.CheckDGMOutput}
\alias{.CheckDGMOutput}
\title{Check Inputs for Data Generating Model Specification}
\usage{
.CheckDGMOutput(dgm_param_vec)
}
\value{
NULL. Function returns NULL if all checks pass.
}
\description{
Internal function to validate the output for the \code{SpecifyDataGeneratingModel} function.
Checks that the output is a named numeric vector.
}
\seealso{
\code{\link{SpecifyDataGeneratingModel}}
}
\keyword{internal,check}
