% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.CreateCrossoverTimePeriodVec}
\alias{.CreateCrossoverTimePeriodVec}
\title{Create an integer vector whose entries are the unique calendar time
periods at which the intervention is introduced to a cluster(s).}
\usage{
.CreateCrossoverTimePeriodVec(n_study_periods, n_clust_trt_seqs)
}
\arguments{
\item{n_study_periods}{The number of study periods}

\item{n_clust_trt_seqs}{The number of randomization sequences}
}
\description{
Tries to guess the calendar time periods at which the intervention is
introduced based on the number of study periods and the number of randomization
sequences.

There are three options currently implemented:
\itemize{
\item If the number of study periods is equal to the number of randomization sequences,
then every time period is in the vector of crossover times.
\item If the number of study periods is one more than the number of randomization sequences,
then every time period except the first is in the vector of crossover times.
\item If the number of study periods is greater than the number of randomization sequences
by more than one, then the crossover times are equally spaced starting in period two.
Throws an error if the number of study periods minus one is not evenly divisible by
the number of randomization sequences.
}
}
