% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.CreateDefaultContrastMat}
\alias{.CreateDefaultContrastMat}
\title{Create a default contrast matrix}
\usage{
.CreateDefaultContrastMat(n_time_model_params, n_trt_model_params)
}
\description{
Create a default contrast matrix
}
