% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_power.R
\name{.CreateLinearPredictors}
\alias{.CreateLinearPredictors}
\title{Create the Vector of linear predictors}
\usage{
.CreateLinearPredictors(design_mat, dgm_param_col_vec)
}
\description{
Create the Vector of linear predictors
}
