% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_power.R
\name{.CreateMeanJacobianMat}
\alias{.CreateMeanJacobianMat}
\title{Create the Mean Jacobian Matrix (\( D_i \))}
\usage{
.CreateMeanJacobianMat(linear_pred_vec, design_mat, family)
}
\arguments{
\item{family}{A family object (from base R \code{stats} class).}

\item{mu_col_vec}{A numeric column vector containing the expected responses \( \mu \).}
}
\value{
Returns a numeric matrix, the mean Jacobian matrix \( D_i \).
}
\description{
This internal function calculates the mean Jacobian matrix \( D_i \), which is the derivative of the expected response \( \mu \)
with respect to the mean parameters \( \beta \), based on the link function.
}
\details{
The mean Jacobian matrix \( D_i \) is calculated based on the link function as follows:
[
D_i = \frac{\partial \mu}{\partial \beta^T}
]

Uses the \code{mu.eta} function from the family object for the calculation
}
\keyword{internal}
