% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/calculate_power.R
\name{.CreateVarParamMatrix}
\alias{.CreateVarParamMatrix}
\title{Create the Variance Parameter Matrix (\( B_i \))}
\usage{
.CreateVarParamMatrix(mu_col_vec, dispersion_scalar, family)
}
\arguments{
\item{mu_col_vec}{A numeric column vector containing the expected responses \( \mu \).}

\item{dispersion_scalar}{A numeric scalar representing the dispersion parameter \( \phi \).}

\item{family}{A family object (from base R \code{stats} class).}
}
\value{
Returns a numeric matrix \( B_i \), the variance parameter matrix.
}
\description{
This internal function calculates the variance parameter matrix \( B_i \) based on the expected responses \( \mu \),
the dispersion parameter \( \phi \), and a variance function \( v \).
}
\details{
The variance parameter matrix \( B_i \) is calculated as:
[
B_i = \phi \, \text{diag}(v(\mu_{i11}), \ldots, v(\mu_{IJK}))
]
where \( \phi \) is the dispersion parameter and \( v(\mu_{ijk}) \) is the variance.

Uses the built in \code{variance} function from the family object for the calculation
}
\keyword{internal}
