% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.NameTimeEffects}
\alias{.NameTimeEffects}
\title{Names for the Time Effects Parameters}
\usage{
.NameTimeEffects(time_model_type, n_study_periods)
}
\description{
Gets its own function because it's used in the model parameter
and design matrix functions.
}
