% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mean_model_param_vector.R
\name{.SpecifyAvgOrLinearTrtEffect}
\alias{.SpecifyAvgOrLinearTrtEffect}
\title{Specify Average or Linear Treatment Effect Parameters}
\usage{
.SpecifyAvgOrLinearTrtEffect(
  mli_study_flag,
  clust_trt_param,
  ind_trt_param,
  interaction_param
)
}
\arguments{
\item{mli_study_flag}{Logical. Flag to indicate if the study is investigating a multi-level intervention.}

\item{clust_trt_param}{Numeric. The treatment effect parameter at the cluster level.}

\item{ind_trt_param}{Numeric. The treatment effect parameter at the individual level. Optional if \code{mli_study_flag} is FALSE.}

\item{interaction_param}{Numeric. The interaction parameter for the treatment effect. Optional if \code{mli_study_flag} is FALSE.}
}
\value{
A named numeric vector with the treatment effect parameters.
}
\description{
Internal function to specify the parameters for average or linear treatment effects in the data generating model.
}
\keyword{internal}
