% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mean_model_param_vector.R
\name{.SpecifyCategoricalTimeEffect}
\alias{.SpecifyCategoricalTimeEffect}
\title{Specify Categorical Time Effect Parameters}
\usage{
.SpecifyCategoricalTimeEffect(n_study_periods, period_effect_param_vec)
}
\arguments{
\item{n_study_periods}{Numeric. The number of study periods, must be at least 2.}

\item{period_effect_param_vec}{Numeric vector. The period effect parameters for the time model.}
}
\value{
A named numeric vector with the time effect parameters.
}
\description{
Internal function to specify the parameters for a categorical time effect in the data generating model.
}
\keyword{internal}
