% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mean_model_param_vector.R
\name{.SpecifyLinearTimeEffect}
\alias{.SpecifyLinearTimeEffect}
\title{Specify Linear Time Effect Parameters}
\usage{
.SpecifyLinearTimeEffect(time_intercept_param, time_trend_param)
}
\arguments{
\item{time_intercept_param}{Numeric. The intercept parameter for the time model.}

\item{time_trend_param}{Numeric. The trend parameter for the time model.}
}
\value{
A named numeric vector with the time effect parameters.
}
\description{
Internal function to specify the parameters for a linear time effect in the data generating model.
}
\keyword{internal}
