% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.WrangleFamily}
\alias{.WrangleFamily}
\title{Input processing for family objects}
\usage{
.WrangleFamily(family)
}
\arguments{
\item{family}{Ideally a family object (list), but if it's a string or function
\code{.WrangleFamily} will attempt to convert it to a family object.}
}
\description{
For ensuring that the input \code{family} is a family object (list)
Handles inputs in the form of a string, function, or list
}
\details{
Description: Utility functions for the project
}
