% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/replicate-prev-sci-example.R
\name{ReplicatePrevSciPaperExample}
\alias{ReplicatePrevSciPaperExample}
\title{Replicate Example Published in Prevention Science}
\usage{
ReplicatePrevSciPaperExample()
}
\description{
Replicates the example published in Prevention Science.
}
