% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_mean_model_param_vector.R
\name{SpecifyDataGeneratingModel}
\alias{SpecifyDataGeneratingModel}
\title{Specify Data Generating Model}
\usage{
SpecifyDataGeneratingModel(
  n_study_periods = NULL,
  mli_study_flag = FALSE,
  time_model_type = "linear",
  time_intercept_param = NULL,
  time_trend_param = NULL,
  period_effect_param_vec = NULL,
  trt_model_type = "linear",
  clust_trt_param,
  ind_trt_param = NULL,
  interaction_param = NULL
)
}
\arguments{
\item{n_study_periods}{Integer. The number of study periods. Must be at least 2.}

\item{mli_study_flag}{Logical. Flag to indicate if the study is investigating a multi-level intervention.}

\item{time_model_type}{Character. The type of time effect model to be used. Valid values are "linear" and "categorical".}

\item{time_intercept_param}{Numeric. The intercept parameter for the time model. Required if \code{time_model_type} is "linear".}

\item{time_trend_param}{Numeric. The trend parameter for the time model. Required if \code{time_model_type} is "linear".}

\item{period_effect_param_vec}{Numeric vector. The period effect parameters for the time model. Required if \code{time_model_type} is "categorical".}

\item{trt_model_type}{Character. The type of treatment effect model to be used. Valid values are "linear" and "average".}

\item{clust_trt_param}{Numeric. The treatment effect parameter at the cluster level.}

\item{ind_trt_param}{Numeric. The treatment effect parameter at the individual level. Only required if \code{mli_study_flag} is TRUE.}

\item{interaction_param}{Numeric. The interaction parameter for the treatment effect. Only required if \code{mli_study_flag} is TRUE.}
}
\value{
A named numeric vector of model parameters.
}
\description{
Wrapper function for specifying the data generating model for a given study.
This function allows for the specification of both calendar time and treatment effects.
}
\examples{
\dontrun{
SpecifyDataGeneratingModel(n_study_periods = 5,
                           mli_study_flag = FALSE,
                           time_model_type = "linear",
                           time_intercept_param = 0.5,
                           time_trend_param = 0.1,
                           trt_model_type = "linear",
                           clust_trt_param = 0.3)
}

}
