% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/swtgeepower-package.R
\docType{package}
\name{swtgeepower-package}
\alias{swtgeepower}
\alias{swtgeepower-package}
\title{swtgeepower: Power and Sample Size Calculation for Stepped Wedge Trials using GEE}
\description{
What the package does (one paragraph).
}
\author{
\strong{Maintainer}: Sperger John \email{first.last@example.com} (\href{https://orcid.org/YOUR-ORCID-ID}{ORCID})

}
\keyword{internal}
