% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/create_correlation_matrices.R,
%   R/create_incidence_matrices.R
\docType{package}
\name{swtgeepower}
\alias{swtgeepower}
\alias{CreateCompleteWorkingCorMatList}
\alias{CreateIncidenceMatricesList}
\title{swtgeepower}
\usage{
CreateCompleteWorkingCorMatList(
  n_study_periods,
  n_clusters,
  n_ind_per_clust,
  sample_type,
  within_period_cor,
  between_period_cor = NULL,
  within_subject_cor = NULL,
  cor_decay_rate = NULL
)

CreateIncidenceMatricesList(n_study_periods, n_obs_periods)
}
\value{
list of incidence matrices
}
\description{
Power and sample size calculations for stepped wedge cluster randomized trials and multilevel cluster randomized trials

r lifecycle::badge("experimental")`
}
\details{
swtgeepower

Power and sample size calculations for stepped wedge cluster randomized trials and multilevel cluster randomized trials
}
